# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ward Civic Issue Portal REST API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: issues
  - name: users
  - name: dashboard
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Start citizen registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registration started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send an OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify an OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpInput"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /auth/logout:
    post:
      operationId: logoutUser
      tags: [auth]
      summary: Log out
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request a password reset OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: Reset OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset a password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
  /issues:
    get:
      operationId: getIssues
      tags: [issues]
      summary: List issues visible to the current user
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: priority
          schema: { type: string }
        - in: query
          name: ward
          schema: { type: string }
      responses:
        "200":
          description: Issues
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Issue" }
    post:
      operationId: createIssue
      tags: [issues]
      summary: Create a civic issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueInput"
      responses:
        "201":
          description: Created issue
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Issue" }
  /issues/{id}:
    get:
      operationId: getIssueById
      tags: [issues]
      summary: Get an issue
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Issue
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Issue" }
    put:
      operationId: updateIssue
      tags: [issues]
      summary: Update an issue as an administrator
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IssueUpdate" }
      responses:
        "200":
          description: Updated issue
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Issue" }
    delete:
      operationId: deleteIssue
      tags: [issues]
      summary: Delete an issue
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Deleted issue
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthMessage" }
  /issues/{id}/status:
    patch:
      operationId: updateIssueStatus
      tags: [issues]
      summary: Update issue status
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StatusUpdate" }
      responses:
        "200":
          description: Updated issue
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Issue" }
  /issues/{id}/priority:
    patch:
      operationId: updateIssuePriority
      tags: [issues]
      summary: Update issue priority
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PriorityUpdate" }
      responses:
        "200":
          description: Updated issue
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Issue" }
  /users/profile:
    get:
      operationId: getProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfileUpdate" }
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /dashboard/citizen:
    get:
      operationId: getCitizenDashboard
      tags: [dashboard]
      summary: Get citizen dashboard summary
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dashboard" }
  /dashboard/admin:
    get:
      operationId: getAdminDashboard
      tags: [dashboard]
      summary: Get admin dashboard summary
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminDashboard" }
components:
  schemas:
    HealthStatus:
      type: object
      properties: { status: { type: string } }
      required: [status]
    AuthMessage:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
      required: [success, message]
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        mobile: { type: string }
        ward: { type: string }
        address: { type: string }
        role: { type: string, enum: [citizen, admin] }
        isVerified: { type: boolean }
      required: [id, name, email, mobile, ward, address, role, isVerified]
    LoginInput:
      type: object
      properties:
        identifier: { type: string, minLength: 3 }
        password: { type: string, minLength: 6 }
      required: [identifier, password]
    LoginResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
      required: [success, message, token, user]
    RegisterInput:
      type: object
      properties:
        name: { type: string, minLength: 2 }
        email: { type: string, format: email }
        mobile: { type: string, minLength: 10 }
        password: { type: string, minLength: 8 }
        ward: { type: string }
        address: { type: string, minLength: 5 }
      required: [name, email, mobile, password, ward, address]
    OtpRequest:
      type: object
      properties:
        email: { type: string, format: email }
        mobile: { type: string }
      required: [email]
    VerifyOtpInput:
      type: object
      properties:
        email: { type: string, format: email }
        otp: { type: string, minLength: 6, maxLength: 6 }
      required: [email, otp]
    ResetPasswordInput:
      type: object
      properties:
        email: { type: string, format: email }
        otp: { type: string, minLength: 6, maxLength: 6 }
        newPassword: { type: string, minLength: 8 }
      required: [email, otp, newPassword]
    Issue:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        category: { type: string }
        description: { type: string }
        ward: { type: string }
        location: { type: string }
        landmark: { type: string }
        priority: { type: string, enum: [Low, Medium, High, Critical] }
        status: { type: string, enum: [Submitted, "Under Review", Assigned, "In Progress", Resolved, Rejected] }
        image: { type: ["string", "null"] }
        adminRemark: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
        citizenName: { type: string }
        citizenEmail: { type: string }
        history:
          type: array
          items: { $ref: "#/components/schemas/StatusHistory" }
      required: [id, title, category, description, ward, location, landmark, priority, status, image, adminRemark, createdAt, updatedAt, citizenName, citizenEmail, history]
    StatusHistory:
      type: object
      properties:
        status: { type: string }
        date: { type: string }
        updatedBy: { type: string }
        remark: { type: ["string", "null"] }
      required: [status, date, updatedBy, remark]
    IssueInput:
      type: object
      properties:
        title: { type: string, minLength: 4 }
        category: { type: string }
        description: { type: string, minLength: 10 }
        ward: { type: string }
        location: { type: string }
        landmark: { type: string }
        priority: { type: string, enum: [Low, Medium, High, Critical] }
        image: { type: ["string", "null"] }
      required: [title, category, description, ward, location, landmark, priority]
    IssueUpdate:
      type: object
      properties:
        status: { type: string }
        priority: { type: string }
        adminRemark: { type: string }
    StatusUpdate:
      type: object
      properties:
        status: { type: string }
        adminRemark: { type: string }
      required: [status]
    PriorityUpdate:
      type: object
      properties:
        priority: { type: string }
      required: [priority]
    ProfileUpdate:
      type: object
      properties:
        name: { type: string }
        mobile: { type: string }
        address: { type: string }
        ward: { type: string }
    Dashboard:
      type: object
      properties:
        total: { type: integer }
        submitted: { type: integer }
        inProgress: { type: integer }
        resolved: { type: integer }
        recent: { type: array, items: { $ref: "#/components/schemas/Issue" } }
      required: [total, submitted, inProgress, resolved, recent]
    AdminDashboard:
      type: object
      properties:
        total: { type: integer }
        submitted: { type: integer }
        underReview: { type: integer }
        assigned: { type: integer }
        inProgress: { type: integer }
        resolved: { type: integer }
        rejected: { type: integer }
        byCategory:
          type: object
          additionalProperties: { type: integer }
      required: [total, submitted, underReview, assigned, inProgress, resolved, rejected, byCategory]