# lib/api-zod/src/generated/types/authMessage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Ward Civic Issue Portal REST API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthMessage {
  success: boolean;
  message: string;
}